@@ -28,7 +28,7 @@ version_tuple: VERSION_TUPLE
 commit_id: COMMIT_ID
 __commit_id__: COMMIT_ID
 
-__version__ = version = '0.2.4.dev6+gc15eee0de.d20251202'
-__version_tuple__ = version_tuple = (0, 2, 4, 'dev6', 'gc15eee0de.d20251202')
+__version__ = version = '0.2.4.dev10+gbbc3626f2.d20260107'
+__version_tuple__ = version_tuple = (0, 2, 4, 'dev10', 'gbbc3626f2.d20260107')
 
-__commit_id__ = commit_id = 'gc15eee0de'
+__commit_id__ = commit_id = 'gbbc3626f2'
