@@ -1,25 +1,26 @@
 """Load and save JSON files easily across files.
 
 Public API:
-- Classes: JSONFile, JsonSerializationSettings
-- Functions: load, sync, reset, close
-- Defaults: DEFAULT_SERIALIZATION_SETTINGS
+    - Classes: JSONFile, JsonSerializationSettings
+    - Functions: load, sync, reset, close
+    - Defaults: DEFAULT_SERIALIZATION_SETTINGS
 """
 
+from .__about__ import __version__
 from .fileutils import (
+    DEFAULT_SERIALIZATION_SETTINGS,
     JSONFile,
     JsonSerializationSettings,
-    DEFAULT_SERIALIZATION_SETTINGS,
 )
-from .pool import load, sync, reset, close
-from .__about__ import __version__
+from .pool import close, load, reset, sync
 
 __all__ = [
-    "JSONFile",
-    "JsonSerializationSettings",
-    "DEFAULT_SERIALIZATION_SETTINGS",
     "load",
+    "DEFAULT_SERIALIZATION_SETTINGS",
     "sync",
+    "JSONFile",
     "reset",
     "close",
+    "JsonSerializationSettings",
+    "__version__",
 ]
