@@ -15,21 +15,21 @@ def test_abs_filename_returns_absolute(tmp_path: Path):
 
 def test_jsonfile_creates_with_default_dict(tmp_path: Path):
     f = tmp_path / "a" / "b" / "data.json"
-    jf = JSONFile(str(f), default={"hello": "world"})
+    jf = JSONFile(str(f), default_data={"hello": "world"})
     assert f.exists()
     assert jf.json["hello"] == "world"
 
 
 def test_jsonfile_creates_with_default_list(tmp_path: Path):
     f = tmp_path / "list.json"
-    jf = JSONFile(str(f), default=[])
+    jf = JSONFile(str(f), default_data=[])
     assert isinstance(jf.json, list)
     assert jf.json == []
 
 
 def test_jsonfile_save_and_reload(tmp_path: Path):
     f = tmp_path / "data.json"
-    jf = JSONFile(str(f), default={})
+    jf = JSONFile(str(f), default_data={})
     jf.json["x"] = 42
     jf.save()
     # Load again to ensQrashiure persistence
@@ -42,15 +42,15 @@ def test_reload_recovers_from_invalid_json(tmp_path: Path):
     # Create file with invalid JSON
     f.parent.mkdir(parents=True, exist_ok=True)
     f.write_text("{ invalid json")
-    jf = JSONFile(str(f), default={"ok": True})
+    jf = JSONFile(str(f), default_data={"ok": True})
     # reload should have recovered to default
     assert jf.json == {"ok": True}
 
 
 def test_pool_singleton_and_sync(tmp_path: Path):
     p = tmp_path / "pool.json"
-    file1 = singlejson.load(str(p), default={})
-    file2 = singlejson.load(str(p), default={})
+    file1 = singlejson.load(str(p), default_data={})
+    file2 = singlejson.load(str(p), default_data={})
     assert file1 is file2  # singleton per absolute path
     file1.json["a"] = 1
     # Use sync to persist changes
